minor fixes and reorganizing include file to eliminate unneeded code generation.
