+++ compiler.cpp
@@ -1,11 +1,10 @@
 #include "scanner.h"
 #include "parser.h"
-#include "macro.h"
-#include "scope.h"
+//#include "macro.h"
+#include "token_type.h"
 #include "scopeTracker.h"
 #include <string>
 #include <iostream>
-#include <cstdio>
 
 using namespace std;
 
@@ -22,38 +21,25 @@ int main(int argc, char **argv){
 		if( (string(argv[1]) == "--help") || (string(argv[1]) == "--h") ){
 			cout << "\nThis is a compiler written for the University of Cincinnati class: EECE6083 Compiler Theory" << endl;
 			cout << "\nThe compiler is an LL(1) recursive descent compiler and utilizes the C++ programming language to scan, parse, and type check the program and LLVM to generate the compiler backend." << endl;
-			cout << "\nTo use this compiler, compile and then run from the command line using the arguments: [ --help | --h | --debug | --d ] filename" << endl;
-			cout << "\nThe compiler will scan and parse your file and generate code if parsing is successful. Otherwise appropriate errors and warnings will be shown" << endl;
+			cout << "\nTo use this compiler, compile and then run from the command line using the arguments: [ --help | --h | --debug | --d ] filename." << endl;
+			cout << "\nThe compiler will scan and parse your file and generate code if parsing is successful. Otherwise relevant errors and warnings will be shown." << endl;
 			cout << "\n--debug or --d argument will print out each token as it is scanned and print out each scope's symbol table after the scope is exited." << endl;
 			return 0;
 		}
 		else if( (string(argv[1]) == "--debug") || (string(argv[1]) == "--d") ){
 			debug = true;
 			filename = string(argv[2]);
-			//test to see if the file can be read
-			ifstream inputfile(filename);
-			if( inputfile.is_open() ) inputfile.close();
-			else{
-				commandError();
-				return 0;
-			}
 		}
-		else{
+		else if(argc == 2){
 			debug = false;
 			filename = string(argv[1]);
-			//test to see if the file can be read
-			ifstream inputfile(filename);
-			if( inputfile.is_open() ) inputfile.close();
-			else{
-				commandError();
-				return 0;
-			}
 		}	
 	}
 	else{
 		commandError();
 		return 0;
 	}
+
 	//Initialize scanner and symbol tables.
 	Scanner *scanner = new Scanner;
 	scopeTracker *scopes = new scopeTracker(debug);
@@ -67,5 +53,5 @@ int main(int argc, char **argv){
 		*token = scanner->getToken();
 		Parser parser(token, scanner, scopes);
 	}
-	else return 0;
+	return 0;
 }
