+++ scanner.h
@@ -3,19 +3,18 @@
 
 #include <string>
 #include <map>
-#include <fstream>
-#include <iostream>
 #include <cstdio>
-#include <stdio.h>
-#include <cstddef>
 #include "macro.h"
+#include "token_type.h"
 
 using namespace std;
 class Scanner
 {
 	private:
 		int line_number;
+		token_type return_token;
 		FILE * fPtr;
+		bool debug;	
 		map<string,int> reserved_table;
 		int ScanOneToken(FILE * fPtr, token_type *token);
 		bool isNum(char character);
@@ -23,15 +22,13 @@ class Scanner
 		bool isString(char character);
 		bool isChar(char character);
 		bool isSingleToken(char character);
-		bool isSpace(char character);
-		bool debug;
+		bool isSpace(char character);	
 	public:
 		Scanner();
 		~Scanner();		
 		bool InitScanner(string filename, bool debug_input);
 		token_type getToken();	
-		void PrintToken();	
-		token_type return_token;
+		void PrintToken();			
 		token_type *token;
 };
